pyproject.toml
[project]
name = "shotstack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mypy>=1.19.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "ruff>=0.15.1",
    "tomli>=2.0.0",
]

[dependency-groups]
dev = [
    "types-requests>=2.32.4.20260107",
]
